@@ -17,7 +17,6 @@
     "test": "ng test",
     "lint": "ng lint && yarn lint:styles",
     "lint:styles": "stylelint \"src/**/*.scss\"",
-    "preinstall": "npx only-allow yarn",
     "postinstall": "ng run demo-angular-material-crud:version",
     "prepare": "husky install",
     "i18n:extract-regex": "ngx-i18n-extract-regex-cli extract --library ngx-translate",
@@ -25,16 +24,16 @@
     "fix-vulnerabilities": "npx yarn-audit-fix && npx browserslist --update-db"
   },
   "dependencies": {
-    "@angular/animations": "^17.0.8",
-    "@angular/cdk": "^17.0.4",
-    "@angular/common": "^17.0.8",
-    "@angular/compiler": "^17.0.8",
-    "@angular/core": "^17.0.8",
-    "@angular/forms": "^17.0.8",
-    "@angular/material": "^17.0.4",
-    "@angular/platform-browser": "^17.0.8",
-    "@angular/platform-browser-dynamic": "^17.0.8",
-    "@angular/router": "^17.0.8",
+    "@angular/animations": "^17.1.2",
+    "@angular/cdk": "^17.1.2",
+    "@angular/common": "^17.1.2",
+    "@angular/compiler": "^17.1.2",
+    "@angular/core": "^17.1.2",
+    "@angular/forms": "^17.1.2",
+    "@angular/material": "^17.1.2",
+    "@angular/platform-browser": "^17.1.2",
+    "@angular/platform-browser-dynamic": "^17.1.2",
+    "@angular/router": "^17.1.2",
     "@gilsdav/ngx-translate-router": "^7.0.0",
     "@gilsdav/ngx-translate-router-http-loader": "^2.0.0",
     "@ngneat/tailwind": "7.0.3",
@@ -49,45 +48,46 @@
     "zone.js": "^0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.9",
-    "@angular-eslint/builder": "^17.1.1",
-    "@angular-eslint/eslint-plugin": "^17.1.1",
-    "@angular-eslint/eslint-plugin-template": "^17.1.1",
-    "@angular-eslint/schematics": "^17.1.1",
-    "@angular-eslint/template-parser": "^17.1.1",
-    "@angular/cli": "^17.0.9",
-    "@angular/compiler-cli": "^17.0.8",
+    "@angular-devkit/build-angular": "^17.1.2",
+    "@angular-eslint/builder": "^17.2.1",
+    "@angular-eslint/eslint-plugin": "^17.2.1",
+    "@angular-eslint/eslint-plugin-template": "^17.2.1",
+    "@angular-eslint/schematics": "^17.2.1",
+    "@angular-eslint/template-parser": "^17.2.1",
+    "@angular/cli": "^17.1.2",
+    "@angular/compiler-cli": "^17.1.2",
     "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
-    "@commitlint/cli": "^18.4.4",
-    "@commitlint/config-conventional": "^18.4.4",
+    "@commitlint/cli": "^18.6.0",
+    "@commitlint/config-conventional": "^18.6.0",
     "@types/jasmine": "^5.1.4",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.10.6",
-    "@typescript-eslint/eslint-plugin": "^6.17.0",
-    "@typescript-eslint/parser": "^6.17.0",
+    "@types/node": "^20.11.16",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
     "@vendure/ngx-translate-extract": "^9.0.3",
-    "autoprefixer": "^10.4.16",
+    "autoprefixer": "^10.4.17",
     "eslint": "^8.56.0",
-    "husky": "^8.0.3",
+    "husky": "^9.0.10",
     "jasmine-core": "^5.1.1",
     "karma": "^6.4.2",
     "karma-chrome-launcher": "^3.2.0",
     "karma-coverage": "^2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "lint-staged": "^15.2.0",
+    "lint-staged": "^15.2.1",
     "ngx-devkit-builders": "^1.2.0",
     "ngx-i18n-extract-regex-cli": "^0.0.6",
     "postcss": "^8.4.33",
-    "prettier": "^3.1.1",
-    "stylelint": "^16.1.0",
+    "prettier": "^3.2.4",
+    "stylelint": "^16.2.1",
     "stylelint-config-standard": "^36.0.0",
     "stylelint-config-standard-scss": "^13.0.0",
     "stylelint-config-tailwindcss": "^0.0.7",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.3"
   },
   "engines": {
-    "node": ">=20 <22"
+    "node": ">=20 <22",
+    "npm": "please-use-yarn"
   }
 }
